requirements.txt
Django==5.1.7
mysqlclient==2.2.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.1
djangorestframework-simplejwt==5.5.0
